# Improve reliability and  performance of resetting ephemeral values