--- a/Sources/StreamChat/Database/DTOs/EphemeralValuesContainer.swift
+++ b/Sources/StreamChat/Database/DTOs/EphemeralValuesContainer.swift
@@ -2,11 +2,15 @@
 // Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
+import CoreData
 import Foundation
 
 /// A protocol marking an DTO object containing ephemeral values, i.e. user online state, or unread counts. These values
 /// need to be reset every time the database is initialized.
-protocol EphemeralValuesContainer {
-    /// Resets the ephemeral values of the container to their default state.
-    func resetEphemeralValues()
+@objc protocol EphemeralValuesContainer {
+    /// Resets the ephemeral relationship values of the container to their default state.
+    static func resetEphemeralRelationshipValues(in context: NSManagedObjectContext)
+    
+    /// Returns batch update request for resetting non-relationship properties.
+    static func resetEphemeralValuesBatchRequests() -> [NSBatchUpdateRequest]
 }