--- a/Sources/StreamChat/Extensions/KeyPath+Extensions.swift
+++ b/Sources/StreamChat/Extensions/KeyPath+Extensions.swift
@@ -1,0 +1,11 @@
+//
+// Copyright © 2024 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+
+extension KeyPath {
+    static func string(_ keyPath: KeyPath) -> String {
+        NSExpression(forKeyPath: keyPath).keyPath
+    }
+}