### 🔗 Issue Links

Resolves: [PBE-6125](https://stream-io.atlassian.net/browse/PBE-6125)

### 🎯 Goal

Improve reliability and performance of resetting ephemeral values by using batch updates and skipping loading DTOs for **every** single object in the DB. There has been reports of crashes when loading all the DTOs in `resetEphemeralValues`.

### 📝 Summary

- Use fetch request for resetting relationship properties (but fetch only what needed)
- Use batch update for resetting non-relationship properties
- Skip fetching every single DTO

### 🛠 Implementation

Remove the convenience protocol for resetting values because it is not efficient and causes fetching all the objects just for resetting values (even when the type does not conform to the protocol). New approach is less flexible but better for performance and reliability.

### 🎨 Showcase

Example from demo app after opening a couple of channels and loading messages for these channels. This is the count of initialized DTOs and time taken by the `resetEphemeralValues`.

| Before | After |
| --- | --- |
| 4525 DTOs | 45 DTOs, 154 with batch update  |
| ~46 ms | ~24 ms  |

### 🧪 Manual Testing Notes

_Explain how this change can be tested manually, if applicable._

### ☑️ Contributor Checklist

- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [ ] This change should be manually QAed
- [x] Changelog is updated with client-facing changes
- [ ] Changelog is updated with new localization keys
- [x] New code is covered by unit tests
- [ ] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (docusaurus, tutorial, CMS)

[PBE-6125]: https://stream-io.atlassian.net/browse/PBE-6125?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ